[HUDI-848] Synchronous commits before completion of instants.
1. Removed code which calls syncTableMetadata
2. Unit tests are broken because not all functionality have been implemented yet.

[HUDI-852] Reader side logic with unit test.

Reader does not need to merge the instants in memory. It simply opens the base and log files, validates which log blocks to read (should have completed instants on dataset timeline).

[HUDI-XXX] Remove option for runtime validation of metadata table against file listing.

Validation does not work in several cases especially with multi-writer. So its best to remove it.

[HUDI-859] Fixed compactions on the Metadata Table.

We cannot perform compaction if there are previous inflight operations on the dataset. This is because a compacted metadata base file at time Tx should represent all the actions on the dataset till time Tx.

[HUDI-878] Support for shards in metadata table partitions.

1. There will be fixed number of shards for each Metadata Table partition.
2. Shards are implemented using filenames of format fileId00ABCD where ABCD is the shard number. This allows easy identification of the files and their order while still keeping the names unique.
3. Shards are pre-allocation during the time of bootstrap.
4. Currently only files partition has 1 shard. But this code is required for record-level-index so implemented here.

[HUDI-850] Do not archive instants on the dataset if they are newer than latest compaction on metadata table.

LogBlocks written to the log file of Metadata Table need to be validated - they are used only if they correspond to a completed action on the dataset.

[HUDI-854, HUDI-855] Handle Metadata Table upgrade/downgrade by deleting the table and re-bootstrapping.

The two versions differ in schema and its complicated to check whether the table is in sync. So its simpler to re-bootstrap as its only the file listing which needs to be re-bootstrapped.

[HUDI-1064] Multi-writer support for metadata table.

Since each operation on metadata table writes to the same files (file-listing partition has a single FileSlice), we can only allow single-writer access to the metadata table. For this, the Transaction Manager is used to lock the table before any updates.

[HUDI-1191] Fixed all unit tests in TestHoodieBackedMetadata.

Fixed checkstyle

Enabling metadata table by default.